# Conflicts:
#	frontend

+++ style/scripts/create_carto.sh
@@ -3,22 +3,23 @@
 DE_VERSION=5.9.0-de7
 EN_VERSION=5.9.0
 
+mkdir osm-carto
+
 if [ "${STYLE:-de}" = 'de' ]; then
-  wget https://github.com/giggls/openstreetmap-carto-de/archive/v${DE_VERSION}.tar.gz
-  tar -xvf v${DE_VERSION}.tar.gz
-  mv openstreetmap-carto-de-${DE_VERSION} osm-carto
+  wget -O- https://github.com/giggls/openstreetmap-carto-de/archive/v${DE_VERSION}.tar.gz | tar -xz -C osm-carto --strip-components=1
+  [ -f "/output/project.mml" ] && cp -f /output/project.mml ./osm-carto/
   echo "Creating osm-de.xml..."
-  sed -i "s/^    dbname: \"gis\".*/    dbname: \"${DBNAME}\"/" ./osm-carto/project.mml
-  carto -a "3.0.22" ./osm-carto/project.mml >/output/osm-de.xml
+  /usr/local/bin/yq -i '._parts.osm2pgsql.dbname = env(DBNAME)' ./osm-carto/project.mml
+  sed -i 's/!!merge <<: /<<: /g' ./osm-carto/project.mml
+  bun run carto -a "3.0.22" ./osm-carto/project.mml >/output/osm-de.xml
   cp -rf ./osm-carto/symbols-de /output/
 else
-  wget https://github.com/gravitystorm/openstreetmap-carto/archive/v${EN_VERSION}.tar.gz
-  tar -xvf v${EN_VERSION}.tar.gz
-  mv openstreetmap-carto-${EN_VERSION} osm-carto
-  cp -f /output/project.mml ./osm-carto/
+  wget -O- https://github.com/gravitystorm/openstreetmap-carto/archive/v${EN_VERSION}.tar.gz | tar -xz -C osm-carto --strip-components=1
+  [ -f "/output/project.mml" ] && cp -f /output/project.mml ./osm-carto/
   echo "Creating osm.xml..."
-  sed -i "s/^    dbname: \"gis\".*/    dbname: \"${DBNAME}\"/" ./osm-carto/project.mml
-  carto -a "3.0.22" ./osm-carto/project.mml >/output/osm.xml
+  /usr/local/bin/yq -i '._parts.osm2pgsql.dbname = env(DBNAME)' ./osm-carto/project.mml
+  sed -i 's/!!merge <<: /<<: /g' ./osm-carto/project.mml
+  bun run carto -a "3.0.22" ./osm-carto/project.mml >/output/osm.xml
   cp -f ./osm-carto/scripts/get-fonts.sh /output/
 fi
 cp -f ./osm-carto/openstreetmap-carto* /output/
